tweaking changes

+++ src/room/DoubleRoom.java
@@ -1,7 +1,7 @@
 package room;
 
 public class DoubleRoom extends Room {
-    protected DoubleRoom(int roomNumber, int price, int capacity, int floor, boolean isOccupied, boolean isSmokingAllowed, boolean hasBalcony) {
+    protected DoubleRoom(int roomNumber, double price, int capacity, int floor, boolean isOccupied, boolean isSmokingAllowed, boolean hasBalcony) {
         super(roomNumber, price, capacity, floor, isOccupied, isSmokingAllowed, hasBalcony);
     }
 }
